+++ CHANGELOG.md
@@ -10,6 +10,7 @@ just to make sure i was not sleeping all the time.
 - added reset filters button
 - added wrapped section
 - fixed markdown rendering
+- added umami analytics
 
 ### 7th April, 2025
 - all applications were added